+++ Gruntfile.js
@@ -97,23 +97,6 @@ module.exports = function(grunt) {
 				}
 			}
 		},
-
-		// @link https://www.npmjs.com/package/grunt-phpdoc
-		phpdoc: {
-			options: {
-				// Task-specific options go here
-				verbose: true
-			},
-			target: {
-				// Target-specific file lists and/or options go here
-				files: {
-					'phpdoc/functions': ['functions/*.php', 'ProgramFunctions/*.php']
-					// 'doc/second': []
-				}
-			},
-			/*src: ['.'],
-			dest: 'phpdoc'*/
-		}
 	});
 
 	/**
@@ -122,7 +105,6 @@ module.exports = function(grunt) {
 	grunt.loadNpmTasks('grunt-contrib-watch');
 	grunt.loadNpmTasks('grunt-contrib-uglify');
 	grunt.loadNpmTasks('grunt-contrib-cssmin');
-	grunt.loadNpmTasks('grunt-phpdoc');
 
 	// Default task(s).
 	grunt.registerTask('default', ['watch']);